Move database service into a devcontainer service
