+++ railties/lib/rails/generators/rails/db/system/change/change_generator.rb
@@ -129,8 +129,8 @@ module Rails
                 compose_config["services"]["rails-app"]["depends_on"]&.delete(database.name)
               end
 
-              if database.service
-                compose_config["services"][database.name] = database.service
+              if database.devcontainer.service
+                compose_config["services"][database.name] = database.devcontainer.service
                 compose_config["volumes"] = { database.volume => nil }.merge(compose_config["volumes"] || {})
                 compose_config["services"]["rails-app"]["depends_on"] = [
                   database.name,
@@ -149,13 +149,13 @@ module Rails
               db_name = database.name
 
               if container_env["DB_HOST"]
-                if database.service
+                if database.devcontainer.service
                   container_env["DB_HOST"] = db_name
                 else
                   container_env.delete("DB_HOST")
                 end
               else
-                if database.service
+                if database.devcontainer.service
                   container_env["DB_HOST"] = db_name
                 end
               end
